Refactor: Update createProject function in utils.ts to include species selection
